@@ -44,7 +44,7 @@ export class FriendIndicator extends React.PureComponent<{}, any> {
         friend_indicator_singleton = this;
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         data.watch(cached.friends, this.updateFriends);
         online_status.event_emitter.on("users-online-updated", this.updateFriendCount);
     }
