@@ -57,7 +57,7 @@ export class StarRating extends React.PureComponent<any, {rating}> {
         }
     }
 
-    componentWillReceiveProps(nextProps) {
+    UNSAFE_componentWillReceiveProps(nextProps) {
         if (this.props.value !== nextProps.value) {
             this.setState({rating: Math.max(0, Math.min(5, nextProps.value))});
         }
