@@ -52,7 +52,7 @@ export class AnnouncementCenter extends React.PureComponent<AnnouncementCenterPr
         };
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         window.document.title = _("Announcement Center");
         this.refresh();
     }
